finalizado primeiro refactor
