+++ src/app.java
@@ -1,0 +1,18 @@
+import interfaces.Formulario;
+
+import java.awt.EventQueue;
+
+public class app {
+	public static void main(String[] args) {
+		EventQueue.invokeLater(new Runnable() {
+			public void run() {
+				try {
+					Formulario frame = new Formulario();
+					frame.setVisible(true);
+				} catch (Exception e) {
+					e.printStackTrace();
+				}
+			}
+		});
+	}
+}
